Refactor: 피드백 반영, 도메인 관점 생각
