--- src/main/java/gift/repository/ProductOptionRepository.java
@@ -1,14 +1,0 @@
-package gift.repository;
-
-import gift.entity.Product;
-import gift.entity.ProductOption;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface ProductOptionRepository extends JpaRepository<ProductOption, Long> {
-    List<ProductOption> findByProductId(Long productId);
-    Boolean existsByProductIdAndName(Long productId, String name);
-}
