+++ src/main/java/gift/service/ProductService.java
@@ -1,9 +1,9 @@
 package gift.service;
 
+import gift.dto.request.ProductOptionRequestDto;
 import gift.dto.request.ProductRequestDto;
+import gift.dto.response.ProductOptionResponseDto;
 import gift.entity.Product;
-import gift.entity.ProductOption;
-import gift.repository.ProductOptionRepository;
 import gift.repository.ProductRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.data.domain.Page;
@@ -17,19 +17,16 @@ import java.util.NoSuchElementException;
 @Transactional
 public class ProductService {
     private final ProductRepository productRepository;
-    private final ProductOptionRepository productOptionRepository;
 
-    public ProductService(ProductRepository productRepository, ProductOptionRepository productOptionRepository) {
+
+    public ProductService(ProductRepository productRepository) {
         this.productRepository = productRepository;
-        this.productOptionRepository = productOptionRepository;
     }
 
     public Product createProduct(ProductRequestDto requestDto) {
         Product product = new Product(null, requestDto.getName(), requestDto.getPrice(), requestDto.getImageUrl());
+        product.addOption(requestDto.getName()+"기본 옵션",9999);
         Product savedProduct = productRepository.save(product);
-
-        ProductOption defaultOption = new ProductOption(requestDto.getName()+"기본 옵션",9999,savedProduct);
-        productOptionRepository.save(defaultOption);
 
         return savedProduct;
     }
@@ -63,4 +60,24 @@ public class ProductService {
         return productRepository.findAll(pageable);
     }
 
+
+
+    public void addOption(Long productId, ProductOptionRequestDto dto) {
+        Product product = productRepository.findById(productId)
+                .orElseThrow(() -> new NoSuchElementException("상품을 찾을 수 없습니다."));
+
+        product.addOption(dto.getName(), dto.getQuantity());
+        // 변경 감지 + cascade 덕분에 ProductOption 자동 저장
+    }
+
+    public List<ProductOptionResponseDto> getOptionsByProductId(Long productId) {
+        Product product= productRepository.findById(productId).orElseThrow(()->new NoSuchElementException("상품을 찾을 수 없습니다"));
+
+        return product.getOptions().stream()
+                .map(ProductOptionResponseDto::from)
+                .toList();
+
+
+
+    }
 }
